+++ BE/baseball-be/src/main/java/kr/codesquad/baseball/player/domain/Player.java
@@ -1,4 +1,4 @@
-package kr.codesquad.baseball.player.repository;
+package kr.codesquad.baseball.player.domain;
 
 import org.springframework.data.annotation.Id;
