+++ BE/baseball-be/src/main/java/kr/codesquad/baseball/team/domain/TeamHasPlayer.java
@@ -1,6 +1,6 @@
-package kr.codesquad.baseball.team.repository;
+package kr.codesquad.baseball.team.domain;
 
-import kr.codesquad.baseball.player.repository.Player;
+import kr.codesquad.baseball.player.domain.Player;
 
 import java.util.List;
 import java.util.stream.Collectors;
